chore(test): extract cluster-related operations to a dedicated utility file Signed-off-by: Anton Misskii <amisskii@redhat.com>
