+++ tests/playwright/src/specs/kind.spec.ts
@@ -16,18 +16,16 @@
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
 
-import { ResourceElementActions } from '../model/core/operations';
-import { ContainerState, ResourceElementState } from '../model/core/states';
 import { ResourceConnectionCardPage } from '../model/pages/resource-connection-card-page';
 import { ResourcesPage } from '../model/pages/resources-page';
-import { VolumesPage } from '../model/pages/volumes-page';
-import { expect as playExpect, test } from '../utility/fixtures';
 import {
+  checkClusterResources,
+  clusterOperations,
   createKindCluster,
-  deleteKindCluster,
-  ensureCliInstalled,
-  getVolumeNameForContainer,
-} from '../utility/operations';
+  deleteCluster,
+} from '../utility/cluster-operations';
+import { expect as playExpect, test } from '../utility/fixtures';
+import { ensureCliInstalled } from '../utility/operations';
 import { waitForPodmanMachineStartup } from '../utility/wait';
 
 const RESOURCE_NAME: string = 'kind';
@@ -51,9 +49,7 @@ test.beforeAll(async ({ runner, page, welcomePage }) => {
 
 test.afterAll(async ({ runner, page }) => {
   try {
-    if (!(process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux')) {
-      await deleteKindCluster(page, KIND_CONTAINER_NAME, CLUSTER_NAME);
-    }
+    await deleteCluster(page, KIND_CONTAINER_NAME, CLUSTER_NAME);
   } finally {
     await runner.close();
   }
@@ -86,7 +82,11 @@ test.describe.serial('Kind End-to-End Tests', { tag: '@k8s_e2e' }, () => {
         await playExpect.poll(async () => resourcesPage.resourceCardIsVisible(RESOURCE_NAME)).toBeTruthy();
       });
     });
-  test.describe('Kind cluster operations', () => {
+  test.describe('Kind cluster validation tests', () => {
+    test.skip(
+      !!process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux',
+      'Tests suite should not run on Linux platform',
+    );
     test('Create a Kind cluster', async ({ page }) => {
       test.setTimeout(CLUSTER_CREATION_TIMEOUT);
       if (process.env.GITHUB_ACTIONS && process.env.RUNNER_OS === 'Linux') {
@@ -96,51 +96,12 @@ test.describe.serial('Kind End-to-End Tests', { tag: '@k8s_e2e' }, () => {
       }
     });
 
-    test('Check resources added with the Kind cluster', async ({ page, navigationBar }) => {
-      const containersPage = await navigationBar.openContainers();
-      await playExpect.poll(async () => containersPage.containerExists(KIND_CONTAINER_NAME)).toBeTruthy();
-      const containerDetailsPage = await containersPage.openContainersDetails(KIND_CONTAINER_NAME);
-      await playExpect.poll(async () => await containerDetailsPage.getState()).toEqual(ContainerState.Running);
-
-      const volumesPage = new VolumesPage(page);
-      const volumeName = await getVolumeNameForContainer(page, KIND_CONTAINER_NAME);
-      if (!volumeName) {
-        throw new Error(`Volume name for container "${KIND_CONTAINER_NAME}" is not defined.`);
-      }
-      const volumeDetailsPage = await volumesPage.openVolumeDetails(volumeName);
-      await playExpect.poll(async () => await volumeDetailsPage.isUsed()).toBeTruthy();
+    test('Check resources added with the Kind cluster', async ({ page }) => {
+      await checkClusterResources(page, KIND_CONTAINER_NAME);
     });
 
-    test('Kind cluster operations - STOP', async ({ navigationBar }) => {
-      await navigationBar.openSettings();
-      await playExpect(kindResourceCard.resourceElementConnectionStatus).toHaveText(ResourceElementState.Running);
-      await kindResourceCard.performConnectionAction(ResourceElementActions.Stop);
-      await playExpect(kindResourceCard.resourceElementConnectionStatus).toHaveText(ResourceElementState.Off, {
-        timeout: 50000,
-      });
-    });
-
-    test('Kind cluster operations - START', async () => {
-      await kindResourceCard.performConnectionAction(ResourceElementActions.Start);
-      await playExpect(kindResourceCard.resourceElementConnectionStatus).toHaveText(ResourceElementState.Running, {
-        timeout: 50000,
-      });
-    });
-
-    test('Kind cluster operatioms - RESTART', async () => {
-      await kindResourceCard.performConnectionAction(ResourceElementActions.Restart);
-      const stopButton = kindResourceCard.resourceElementConnectionActions.getByRole('button', {
-        name: ResourceElementActions.Stop,
-        exact: true,
-      });
-      await playExpect(stopButton).toBeEnabled({ timeout: 25_000 });
-      await playExpect(kindResourceCard.resourceElementConnectionStatus).toHaveText(ResourceElementState.Running, {
-        timeout: 50000,
-      });
-    });
-
-    test('Kind cluster operations - DELETE', async ({ page }) => {
-      await deleteKindCluster(page, KIND_CONTAINER_NAME, CLUSTER_NAME);
+    test('Kind cluster operations', async ({ page }) => {
+      await clusterOperations(page, kindResourceCard, KIND_CONTAINER_NAME, CLUSTER_NAME);
     });
   });
 });
