+++ tests/playwright/src/utility/operations.ts
@@ -23,15 +23,12 @@ import test, { expect as playExpect } from '@playwright/test';
 
 import { ResourceElementActions } from '../model/core/operations';
 import { ResourceElementState } from '../model/core/states';
-import type { KindClusterOptions } from '../model/core/types';
 import { CLIToolsPage } from '../model/pages/cli-tools-page';
-import { CreateKindClusterPage } from '../model/pages/create-kind-cluster-page';
 import { RegistriesPage } from '../model/pages/registries-page';
 import { ResourceConnectionCardPage } from '../model/pages/resource-connection-card-page';
 import { ResourcesPage } from '../model/pages/resources-page';
 import { VolumeDetailsPage } from '../model/pages/volume-details-page';
 import { NavigationBar } from '../model/workbench/navigation';
-import { StatusBar } from '../model/workbench/status-bar';
 import { waitUntil, waitWhile } from './wait';
 
 /**
@@ -310,98 +307,6 @@ export async function ensureCliInstalled(page: Page, resourceName: string, timeo
     await playExpect
       .poll(async () => await cliToolsPage.getCurrentToolVersion(resourceName), {
         timeout: timeout,
-      })
-      .toBeTruthy();
-  });
-}
-
-export async function createKindCluster(
-  page: Page,
-  clusterName: string,
-  usedefaultOptions: boolean,
-  timeout: number = 300_000,
-  { providerType, httpPort, httpsPort, useIngressController, containerImage }: KindClusterOptions = {},
-): Promise<void> {
-  return test.step('Create Kind cluster', async () => {
-    const navigationBar = new NavigationBar(page);
-    const statusBar = new StatusBar(page);
-    const kindResourceCard = new ResourceConnectionCardPage(page, 'kind', clusterName);
-    const createKindClusterPage = new CreateKindClusterPage(page);
-
-    const settingsPage = await navigationBar.openSettings();
-    const resourcesPage = await settingsPage.openTabPage(ResourcesPage);
-    await playExpect(resourcesPage.heading).toBeVisible({ timeout: 10_000 });
-    await playExpect.poll(async () => resourcesPage.resourceCardIsVisible('kind')).toBeTruthy();
-    await playExpect(kindResourceCard.createButton).toBeVisible();
-
-    if (await kindResourceCard.doesResourceElementExist()) {
-      console.log(`Kind cluster [${clusterName}] already present, skipping creation.`);
-      return;
-    }
-
-    await kindResourceCard.createButton.click();
-    if (usedefaultOptions) {
-      await createKindClusterPage.createClusterDefault(clusterName, timeout);
-    } else {
-      await createKindClusterPage.createClusterParametrized(
-        clusterName,
-        {
-          providerType: providerType,
-          httpPort: httpPort,
-          httpsPort: httpsPort,
-          useIngressController: useIngressController,
-          containerImage: containerImage,
-        },
-        timeout,
-      );
-    }
-    await playExpect(kindResourceCard.resourceElement).toBeVisible();
-    await playExpect(kindResourceCard.resourceElementConnectionStatus).toHaveText(ResourceElementState.Running, {
-      timeout: 15_000,
-    });
-    await statusBar.validateKubernetesContext(`kind-${clusterName}`);
-  });
-}
-
-export async function deleteKindCluster(
-  page: Page,
-  containerName: string = 'kind-cluster-control-plane',
-  clusterName: string = 'kind-cluster',
-): Promise<void> {
-  return test.step('Delete Kind cluster', async () => {
-    const navigationBar = new NavigationBar(page);
-    const kindResourceCard = new ResourceConnectionCardPage(page, 'kind', clusterName);
-    const volumeName = await getVolumeNameForContainer(page, containerName);
-
-    await navigationBar.openSettings();
-    const resourcesPage = new ResourcesPage(page);
-    await playExpect(resourcesPage.heading).toBeVisible({ timeout: 10_000 });
-    if (!(await kindResourceCard.doesResourceElementExist())) {
-      console.log(`Kind cluster [${clusterName}] not present, skipping deletion.`);
-      return;
-    }
-
-    await kindResourceCard.performConnectionAction(ResourceElementActions.Stop);
-    await playExpect(kindResourceCard.resourceElementConnectionStatus).toHaveText(ResourceElementState.Off, {
-      timeout: 50_000,
-    });
-    await kindResourceCard.performConnectionAction(ResourceElementActions.Delete);
-    await playExpect(kindResourceCard.markdownContent).toBeVisible({
-      timeout: 50_000,
-    });
-    const containersPage = await navigationBar.openContainers();
-    await playExpect(containersPage.heading).toBeVisible();
-    await playExpect
-      .poll(async () => containersPage.containerExists(containerName), {
-        timeout: 10_000,
-      })
-      .toBeFalsy();
-
-    const volumePage = await navigationBar.openVolumes();
-    await playExpect(volumePage.heading).toBeVisible();
-    await playExpect
-      .poll(async () => await volumePage.waitForVolumeDelete(volumeName), {
-        timeout: 20_000,
       })
       .toBeTruthy();
   });
